Return to inventory from shopping cart
In Data.js, I first added ids to my inventory to be able to search through it easier by the product id. In Cart.js, I added a for loop to be able to look at the ids and if an id matches up it adds it back to the inventory.